Switch to OZ AccessControl for Bread.sol (#4)
The main benefit here is the ability to have multiple signers. An example of a future use case where this might come in handy is allowing a trusted external contract to add credit to a customer's account.

Separating what is effectively super admin and admin functionality is also a nice security feature.